Moved delegate to common/ from core/.
