+++ src/common/delegate.js
