+++ src/core/entity.js
@@ -4,7 +4,7 @@
 
 define( function ( require ) {
 
-    var Delegate = require( './delegate' );
+    var Delegate = require( 'common/delegate' );
 
     /* Entity
      *
