+++ src/graphics/component/model.js
@@ -8,7 +8,7 @@ define( function ( require ) {
 
     var math = engine.math;
     var Component = require( 'base/component' );
-    var Delegate = require( '../../core/delegate' );
+    var Delegate = require( 'common/delegate' );
     var Material = require( '../resource/material' );
     var Mesh = require( '../resource/mesh' );
